feat(sync-service): Scale down database connections on replication idle timeout (#3257)
This PR supersedes https://github.com/electric-sql/electric/pull/3126
because the latter has accumulated a bunch of out-of-scope context
making it difficult to make out the feature requirements and follow the
implementation.

Closes #3259

The premise of scaling down database connections is letting the database
itself scale down and release compute resources when it is idle, i.e. no
reads and no writes are occurring. That would be typical of a dev
database where during the night there's no load on the database and no
shape requests to wake it up.

A new configuration option name `ELECTRIC_REPLICATION_IDLE_TIMEOUT` is
introduced: if no new transactions are streamed within that time
interval, Electric will close all database connections until the next
shape request arrives. By default, Electric keeps database connections
open forever.

Remaining tasks:
- [x] Finish the integration test
- [x] Add code comments for new functionality where it's missing
- [x] Prevent StackTelemetry from waking up the connection subsystem
during its periodic WAL size measurement

Addressing concerns from the previous PR's review:

@icehaunter 
> I'm worried that in absence of a replication connection we might cause
WAL to grow indefinitely on the PG.

In a dev environment WAL buildup is rarely a problem. When the developer
is not actively using the application, there will be no database writes
and no shape requests. If there are shape requests, the replication
connection will be running and no WAL buildup will be occuring. We could
come up with an edge case where a script is running in background to
simulate writes to the database and that could lead to WAL buildup once
Electric closes the replication connection. But this is the
responsibility of the developer: they are in full control with regards
to setting the value for `ELECTRIC_REPLICATION_IDLE_TIMEOUT`. It's kind
of similar to the situation where the same background script is
generating writes to the database while Electric itself is not running
and therefore not advancing the replication slot.

In a scenario where Electric has shapes only for some tables that are
rarely written to while other tables are seeing regular writes, the
concern is that once Electric closes its replication connection, those
other writes that are invisible to Electric will cause WAL buildup. In
this case, though, the developer shouldn't set
`ELECTRIC_REPLICATION_IDLE_TIMEOUT` at all, since writes to non-shape
tables will prevent the database server itself from scaling to zero
anyway.

@msfstef 
> rather than waking up on _every_ request, would it not be better if we
did the following:
> 
> 1. on new shape requests, recover db connections (and hold the
requests until active)
> 
> 2. on live shape requests, perform (throttled) checks of the WAL to
decide whether we ned to wake up db conns
> 
>    3. with a relatively long interval, check the WAL for build up
> 
> 
> for 2 and 3 we can have a single WAL checker process that handles
things


Moving the ShapeStatusOwner up the supervision tree has brought us a
step closer to the ability to serve existing shape requests without
waking up the connection subsystem. Another thing we'd need for this is
https://github.com/electric-sql/stratovolt/issues/810. However, I think
triaging requests to see which ones can be served without the database
is a case of premature optimization.

Think about the primary use case for scale down: introducing any sort of
delay (including via throttling) between a write getting committed to
the database and that same write being served to live shape requests
will make it appear as though Electric is not working correctly. And by
nature of Electric's protocol, the vast majority of shape requests will
be issued in a chain until reaching the up-to-date offset, at which
point they transition to live requests. So we inevitably end up in the
situation where the connection subsystem needs to be woken up to serve
live shape requests without artificial lag.

---------

Co-authored-by: Ilia Borovitinov <ilia@electric-sql.com>