feat: Release live SSE mode from experimental to available (#3267)
I've added a patch update for both client and server but this could be a
minor as this is adding a new feature.

I'm slightly confused as to how errors are being handled in SSE mode, as
there's a reference to a flag that does not exist as well. @balegas not
sure how well tested this is especially with the various errors, garbage
collections, timeouts that can occur for shapes?

This also contains a fix for our must refetch messages that were always
being sent as a list of messages even when `in_sse?` was true outside of
the live sse clause, which breaks old clients, so the new server will
now not be compatible with old clients :/ @balegas what do you reckon is
the best thing to do here - IMO since this was an experimental feature
we should be fine with breaking old clients using it before it was
marked as released.


## UPDATE

We are no longer deprecating the experimental flag, we are completely
removing it, so now this update requires anyone using SSE to upgrade
both their server and clients.

## UPDATE 2

We are no longer sending the `must-refetch` message over the SSE event
stream, instead letting it come through a regular 409 response that also
contains the new shape handle in the headers and processed by the client
the same way.

This maintains backwards compatibility, and thus I reinstated the
deprecations of the experimental flags and changed back the changeset
message to not be a breaking one (although still a minor release).

Electric has a test ensuring the the `experimental_live_sse` flag still
works and can thus serve both old and new client versions, and the
client always sends both to ensure it works with both old and new server
versions.

@KyleAMathews @balegas let me know if you agree with this.