feat(sync-service): repopulate log collectors state from the status ets (#3242)
Don't rely on consumer processes existing in order to initialise the
shape log collector -- all the information we need is in the shape
status ets.

We still need the consumer processes to register in order to synchronise
the filters etc with the existing shapes but we only need to do that if
the consumer was not in the original state.

This is handled by having an extra `action` param to the subscribe call.
Shape consumers being restored don't add their shape to the shape log
collector's internal filters while those being newly created do.