feat: Restore `PublicationManager` state from `ShapeStatus` ETS on startup (#3275)
## Edited by @msfstef 

Closes https://github.com/electric-sql/electric/issues/3245

1. On bootup, pub man filters state is recovered directly from
`ShapeStatus`
2. It immediately goes into a `:continue` clause to update the
configuration
3. Keeps track of a `restore_complete?` flag as well as
`restore_waiters` from the new `wait_for_restore` sync API
3a. I decided to add waiters as I want the update to be able to fail due
to e.g. network errors and still accept requests and handle them without
crashing the shape subsystem, in the spirit of previous work.
4a. If restore succeeds, we reply to the waiters
4b. If restore fails, we go into a refresh loop with a shorter timeout
than our regular refresh timeout until restore completes
5. We `wait_for_restore` in the `ShapeCache` before we mark consumers as
ready, meaning we block there until we manage to recover the
publication.

The last step is still coupling and blocking ShapeCache to PubMan but in
the spirit of maintaining the current behaviour but avoiding the
onslaught of recovery messages from all shapes I think this is a good
first step, and @magnetised feel free to move things around in your work
to decouple things.

The main thing to note is that _we should not_ start streaming changes
until we know for sure that the publication has actually been restored -
but I believe the shape subsystem should be able to handle existing
shape data while we wait for that to happen.

I've also made it so that we only call `add_shape` to the publication
manager if a snapshot _has not_ been started, so that most restored
shapes don't end up calling the pubman (@magnetised was this the
intention? if not, I can let you sort it out in your own PRs)

---------

Co-authored-by: msfstef <msfstef@gmail.com>