fix: Fix publication update deadlocks (#3285)
We observed deadlocks during publication updates in high traffic usage.

The protection against concurrent transactions causing deadlocks was
removed here: https://github.com/electric-sql/electric/pull/3218

The idea was that since we linearise everything behind the publication
manager, this protection is redundant, but it seems this can still occur
(and DDL operations on Postgres have much weaker guarantees).

I've added a slow test (tagged slow to only run on CI) that managed to
reproduce the deadlocks, although not always but that's the problem with
concurrency bugs, and restoring the logic of updating replica identities
only after touching the publication seems to have done away with the
error completely.