chore(sync-service): Recover shapes without holding supervisor up (#3293)
This is a small part of the shape subsystem work. This PR recovers
shapes without holding it's supervisor up. Eventually this will mean we
can recover shapes asynchronously while the database connection is being
established.

The effect of this currently is that the ExpiryManager will start before
the shapes have been restores so this PR also makes the ExpiryManager
check if the stack is ready before expiring.

I've also used this opportunity to re-write the ExpiryManager tests that
were a mess (I wrote them)

The other process this affects is the SchemaReconciler, it may now
reconcile before the shapes have been restored but given that it loads
its own data from PersistentKV I don't think this is an issue (but
please let me know if you know otherwise!)