chore(sync-service): Add MonitoredCoreSupervisor (#3307)
Part of a series of PRs to separate the shape and connection subsystems.

This PR moves the StatusMonitor further up the supervision tree so that
it can supervise both subsystems. Currently the StatusMonitor is
supervised by the Connection.Supervisor which means once the shape
subsystem has been moved out of the Connection.Supervisor's descendants
it will no longer be at the right level.

This PR introduces the idea of the electric core - the shape and
connection subsystems, supervised by the CoreSupervisor. There is also a
MonitoredCoreSupervisor which supervises the CoreSupervisor along with
the StatusMonitor with the the rest_for_one strategy which allows the
StatusMonitor to stay consistent (everything it monitors is restarted if
it dies, so there is no chance it will die and lose references to what
it is monitoring).

The supervision tree now looks like this:

StackSupervisor
- utility processes such as the EtsInspector that can be restarted
indecently
- MonitoredCoreSupervisor
  - StatusMonitor
  - CoreSupervisor
- (to be added in a later PR) Replication.Supervisor (to later be
renamed to ShapeSubsystemSupervisor)
- Connection.Manager.Supervisor (to later be renamed to
ConnectionSubsystemSupervisor)