feat: Support generated columns in Postgres 18 (#3297)
Closes https://github.com/electric-sql/electric/issues/3220

## Overview of changes

I've made the following core changes to support this:

- When creating a new publication, if on PG18 and above, always set
`publish_generated_columns = 'stored'`, see [PG
docs](https://www.postgresql.org/docs/current/sql-createpublication.html#SQL-CREATEPUBLICATION-PARAMS-WITH-PUBLISH-GENERATED-COLUMNS)
for details
- If we find an existing publication, whether owned or not, we do not
alter it.
- For people upgrading Electric on a PG18 db, they will have to manually
set the publication parameter or somehow reset the electric
slot/publication/timeline to get this feature. I could add a step to
always enable it if we can, but ideally in a separate PR.
- In the `PublicationManager`, before each update, we check the
publication status which now includes whether generated columns are
replicated
- We cache the result, and we initialise with an optimistic assumption
that generated columns are supported
- In PubMan, we now keep track of which shape handles use generated
columns, and in case the configuration changes we do an O(N) scan over
our tracked shapes and schedule removals for any shape that used
generated columns
- Slow scan is fine, it's just a list traversal that happens _only_ if
the publication configuration is disabled at runtime, which might be the
case for manually configured publications but never for our own owned
publication.
- Any shapes pending creation that require generated columns would also
receive an error reply
- Once the configuration result is cached, shape creations immediately
get an error answer from the publication manager until a check confirms
that generated columns are supported
- The Inspector now loads `supported_features` from Postgres, so at the
validation stage we can determine if we should block generated column
shapes the same way we have until now, or whether we should let them
through and let the PublicationManager determine if they can be created.
- We will now get different errors if you are on e.g. PG17 and request a
shape with generated columns, and if you are on PG18 and above. In the
latter you now receive:
  ```
Publication "{pub_name}" does not publish generated columns. This is a
feature introduced in PostgreSQL 18 and requires setting the publication
parameter 'publish_generated_columns' to 'stored'. Alternatively, you
can exclude them from the shape by explicitly listing which columns to
fetch in the 'columns' query param.
  ```
- Refactored the Inspector ETS table name to `inspector_table` rather
than `relation_table` for more generality as it now holds more Postgres
inspection information - I think it makes sense to reuse the Inspector
for this to keep the infrastructure decoupled, this is purely
information about what Postgres supports as part of a validation step.
- Parse PublicationManager error, which gets turned into a
`SnapshotError`, at the API level to return 400 in case the feature is
supported in the database but not configured.
- Some minor typespec fixes and cleanups of unnecessary info (e.g.
`can_alter_publication?` flag in ConnMan - the pubman will determine it
by itself at runtime)


## Potential alternatives
- Rather than return errors from the publication manager, we can just
clean up shapes when the configuration changes and set the flag
appropriately somewhere so that we continue to reject generated columns
in the same way at the shape creation validation stage
- PRO: simplify handling of cases by introducing a sort of "supported
features" table for the validation to lookup
- PRO: by separating into a separate "supported features" table we could
offload it from the inspector (although it would still need to check
postgres somehow, so I don't like this)
- CON: I think it's important to separate when a feature is not
available and when it is not enabled by choice - by letting the
publication manager reply, we are notifying developers that they _could_
use generated columns if they wanted to if they are using PG18 and
above.