fix: fix tanstack-db-web-starter dependency leak by reformatting project to use quotes instead of backticks
In a previous commit an eslint rule was introduced, which forced the use of backticks for every string. As this caused issue in the tanstack-db-web-starter, this commit reverts this and instead adds the two eslint rules [prefer-template](https://eslint.org/docs/latest/rules/prefer-template) and [no-template-curly-in-string](https://eslint.org/docs/latest/rules/no-template-curly-in-string), which should encourage the usage of template strings, while not breaking any examples.
This commit also adds .idea to the .gitignore as it was missing until now and shouldn't be commited to vsc.