fix: Force stack restart in case of fatal publication misconfiguration (#3324)
Fixes https://github.com/electric-sql/electric/issues/3322

This addresses issues raised in
https://github.com/electric-sql/electric/pull/3321 and will fix the
corresponding integration test once this is merged as well.

I've renamed `restart_connection_subsystem` to
`restore_connection_subsystem` as it only does something if it is
"sleeping", and added a call API for restarting to actually restart the
subsystem. I don't think this is abusing the restarter as I think it is
the right layer of indirection to perform this operation.

We could in the future be more precise with how we do this recovery but
I think for these fatal misconfigurations this is good enough.

---------

Co-authored-by: Oleksii Sholik <oleksii@sholik.dev>