Remove explicit calls to :erlang.garbage_collect() (#3326)
The code is not sitting still and at least one of these calls is
becoming obsolete with Gary's work on starting consumer processes
lazily. Moreover, the current startup of consumer processes in
ShapeCache already spawns a separate `Task` process for each consumer,
so it's not that clear where all the garbage is coming from.

In any case, we should refrain from adding explicit GC unless we can
reliably reproduce the situtations the show its unquestionable benefit.