fix: Fix relation ref counting in publication manager (#3327)
We were removing a relation from prepared relations manually rather than
waiting for all appropriate shape removals, causing issues in checks we
do with ref counting.

We should rely solely on clients removing themselves as we do in the
rest of the publication manager code and rely on the ref counting.