feat(sync-service) Gracefully handle no connection while processing txn in ShapeLogCollector (#3328)
Part of the shape subsystem work:

This PR gracefully handles connection issues while processing
transactions in the ShapeLogCollector. It simply returns a connection
error which the ReplicationClient responds to by waiting for the
connection to be re-established and then retrying.

Please note that the ShapeLogCollector will still crash if there is no
connection while subscribing, and will be shut down by the connection
manager if the replication client losing connection. These scenarios
will be handled by later PRs.