fix: Handle more inspector connection failures (#3320)
Fixes https://github.com/electric-sql/electric/issues/3316

This PR also semi-addresses [this Sentry
issue](https://electricsql-04.sentry.io/issues/62065592/?environment=production&project=4508410462404688&query=is%3Aunresolved&referrer=issue-stream)
- this is again a symptom of in-flight requests during a stack shutdown,
they will still fail but with less misleading errors.


In our shape validation steps, the `:connection_not_available` errors
are handled more gracefully but with a lot of code. This time in the
schema header there wasn't an easy way to thread things through, but we
should be able to just raise appropriate errors and let the error
handler return the appropriate 503 errors.

We do indeed do that, but only for `DBConnection.ConnectionError` -
perhaps we should rethink it a bit.

I've also considered refactoring the code in the shape validation
handling `:connection_not_available` gracefully and instead just raising
`DBConnection.ConnectionError` so that a 503 is returned. Would like to
hear the reviewer's thoughts.

This particular fix is hard to put in a test cause it requires losing
both the cache and connection after a request has been validated and
processed but before the schema header is put in - I could do it with
Repatch but decided against adding flimsy tests and just make it make
sense, I've tested it manually ofc.

p.s. I also took the opportunity to fix the flake in the nullability
router integration test