feat: Move connection lock inside replication connection (#3319)
Closes https://github.com/electric-sql/stratovolt/issues/811

Removes `LockConnection` as a separate connection process altogether,
reducing the number of wal sender processes required to run electric
from 2 to 1 and simplifying the connection manager further.

## Overview of changes
- Acquiring the lock is now part of our regular `ConnectionSetup` state
machine that runs on the replication connection
- System identification is also now part of the regular state machine
setup
- The backend pid is queried along with other `pg_info` to reduce number
of separate queries
- Lock acquisition query failures now immediately retry (removed backoff
logic internal to LockConnection)
- I had added backoff logic but having investigated logs of the past 2
months on cloud we never fail the query for any reason other than
statement timeouts, which should immediately be retried.
- Other failures will trigger a full retry and reestablishing the
connection, like any connection failure would do
- On `terminate` of the connection manager, if the admin pool is
available we forcibly kill the replication conn backend (which includes
both the lock and the replication slot usage)
- Since killing the replication client also loses the lock, we cannot
guarantee that `drop_slot` runs while we hold the lock
- This was not a guarantee before either, so I decided to not concern
myself too much with this undocumented feature anyway - it works with
how we use it internally in cloud
- I ensure the publication is dropped while we hold the lock such that
anyone who grabs the lock will be forced to recreate the slot anyway, as
well as ensuring there's no unnecessary buildup


## Followup work
I believe that we should not rely on the admin pool to do cleanups like
killing the replication client backend and dropping the slot, since the
admin pool is only started after we do initial configuration of the
replication connection which includes the lock and replication
slots/publication.

We should be performing these cleanups with a one-off, single conn pool
connection so we can always at least attempt them upon shutdown.

Ensuring the "drop slot on shutdown" is done in a "protected" way would
require us doing a dance with multiple locks like the one I intend to do
for https://github.com/electric-sql/electric/issues/3286 - grab a
cleanup transaction lock, stop the replication client (and lose the
general lock), do the clenaup, release the transaction lock. Other
electrics check for the cleanup lock before setting up publication/slot
to avoid overlaps. All of this might also be overkill as we cannot
guarantee cleanup of a slot in case of unexpected kills/shutdowns.


## Notes on ConnMan
With this change, conn man has become so simple that I think we're very
close to just using a regular supervisor, which will also allow us to
rethink how we want to do cleanups as we can have a separate monitor
process that takes care of these cleanups with its own ephemeral
connection (if able to establish it).