feat(sync-service): Gracefully handle connection not available in subscribe and relation changes (#3333)
Part of the shape subsystem work:

This PR gracefully handles connection issues while:
- processing relations in the ShapeLogCollector. It simply returns a
connection error which the ReplicationClient responds to by waiting for
the connection to be re-established and then retrying.
- processing shape consumer subscriptions. The shape consumer responds
by shutting down and removing the shape.

Please note that the ShapeLogCollector will still be shut down by the
connection manager if the replication client losing connection. This
scenarios will be handled by later PR.