chore(sync-service): Remove `consumers_ready` notification (#3343)
Part of a series of PRs to separate the shape and connection subsystems.

This PR removes the `consumers_ready` notification which is sent by the
ShapeCache to signal that all the shapes have been restored. Once the
shape subsystem is separated from the connection subsystem,
Connection.Manager will restart without the ShapeCache restarting and so
the `consumers_ready` notification cannot be sent. Instead
Connection.Manager just starts the ReplicationClient.

There's a couple of race conditions to consider here:
1. The replication client may start sending transactions to the
ShapeLogCollector before it's ready, but this is fine because the
ShapeLogCollector will return not_ready and the ReplicationClient will
wait for the stack to be ready then retry.
2. An API request comes in after the shapes have restored but before the
integrity checks (timeline checks) have occurred. This can't currently
happen but once the connection subsystem restarts without restarting the
shape subsystem it will be able to happen. This PR introduces an
additional check on the status monitor so that the stack is not
considered ready until the integrity checks have completed. This ensures
that a request coming in doesn't access shapes that are no longer valid.