feat(sync-service): Make shape subsystem resilient to connection failures (#3344)
This PR allows the shape subsystem (processes under the soon to be
renamed Replication.Supervisor) to stay running even when the connection
subsystem is torn down due to a connection failure.

I've run a bit of chaos testing, dropping the connection randomly while
creating shapes. We still get 500s for the snapshot connection failure
which I'll address on a separate PR, but otherwise it looks good.