feat: Ensure `PublicationManager` handles relations individually and concurrently with requests (#3359)
This work was prompted by timeouts and errors observed on stacks with a
high number of relations involved.

As soon as the sync service was up, many requests would come in for all
relations, or a recovery would be attempted, and we would attempt to
configure all relations in a single transaction.

That would occasionally time out, but also we would often see some
tables being slow to update because e.g. one particular table was
holding the lock perhaps due to db administration occurring, leading to
_no_ tables being configured.

In this PR I am converting PublicationManager to a supervisor of two
processes, a `RelationTracker` and a `Configurator`.

## `RelationTracker`

Responsible for reference counting and request handling. This never
blocks and is always able to reply to shapes wanting to be registered or
deregistered from the publication.

It keeps track of `prepared_relation_filters`,
`submitted_relation_filters`, `committed_relation_filters`, which
represent the three possible states of a relation.

When a relation needs to be in the publication, it will be in the
`prepared` set due to the reference counting logic.

If the `prepared` set differs from the `submitted` or `committed` set,
then a publication update is requested asynchronously from the
`Configurator`. As soon as the update is requested, the `submitted` set
is set to be the `prepared` set.

As the `Configurator` configures relations individually, it sends back
results of whether it succeeds or fails. Any relation that has
successfully configured gets added to the `committed` set.

The `RelationTracker` will immediately reply to requests to add a
relation when the relevant relation is in both the `submitted` and
`committed` set - as in this case we know that the relation has both
been committed and is maintained in any current "in flight" requests to
update it.


## `Configurator`

I have essentially destructured the logic in the
`Electric.Postgres.Configuration` module so that we run each relation
individually in the `Configurator` process.

We first check the publication status, returning any errors or relevant
states like generated column support back to `RelationTracker`.

Then we determine "relation actions" by listing relations in the
publication, replica identities, any invalid relations, etc.

Finally we schedule each action to run via `handle_continue` clauses
terminating when actions are done. We still don't process any messages
while this is happening, but this process only does the publication
update so it should not be a problem.


## `PublicationManager`

This is now a supervisor with a `:one_for_one` strategy - if the
`Configurator` dies, we can just bring it back up again as it holds no
important state other than caching some properties of the publication
and it pulls the filters from the `RelationTracker`. The
`RelationTracker` holds the reference counting but it restores it from
the in memory shape status cache, so in case of a restart it will get
back to where it is, and it can handle updates it does not know about
from the configurator.

Important to note that I've also reinstated all shapes asking the
publication to be added regardless of whether they have a snapshot as we
are not lazily loading shapes and it is not longer a problem doing this.