fix: Parse disk full and duplicate slot file errors (#3378)
Fixes sentry errors
[#1](https://electricsql-04.sentry.io/issues/74727274/) and
[#2](https://electricsql-04.sentry.io/issues/74727257)

I have marked the `:disk_full` error as retryable, since storage might
be freed up automatically or added by the db administrator as a response
to this error and should thus not shut down the system.

I have marked the `:duplicate_file` for the replication slot
specifically as retryable as well, as it is a tmp file for an atomic
write that seems like a race. Interesting that this occurred and might
be worth looking into if it keeps occurring.