feat(sync-service): Speed up shape meta data removal (#3385)
For 50K shapes and concurrent read and writes of the ShapeStatus ETS
tables, `ShapeStatus.remove_shape` takes 50ms on Ampere, 23ms locally.
This only gets slower as the number of shapes increases. This PR speeds
up that process by 450% to 0.05ms.

It does mean that the metadata delete is not atomic, but as far as I can
tell there are no race conditions that this introduces.