feat: Instrument publication manager updates (#3388)
With the publication manager updates I merged last week I removed the
`publication_manager.update_publication` span as we no longer do a
single update for all relations - but did not add back the appropriate
telemetry.

This PR adds the span `publication_manager.update_publication_relation`
which is a span per relation update, along with metadata on what the
action was (`add`, `add_and_configure`, `configure`, `drop`) and the
relation name (without OID). I think it needs to be a different name as
it fundamentally is a different operation.

This includes minor refactoring to avoid duplicating telemetry code -
and reduces the use of `:continue` clauses.