chore: Remove storage metadata backup logic (#3400)
Now that lazy shape loading is in place, backing up and reloading
storage metadata in the already overloaded ShapeStatus table does not
offer any significant benefits.

The only reason it was there was because on startup if we had lots of
shapes to recover, we would have thousands of FS operations queued up
for reading lots of different metadata from files, and that would block
recovery and the ability of the system to start serving requests.

With lazy shape loading, we no longer have an onslaught of FS operations
blocking the system recovery, instead they occur on demand and
proportionally to shapes requested after a reboot. Holding the cache in
memory until a shape is loaded feels unnecessary and one more thing that
can go wrong.