chore(sync-service): Reduce memory consumption of consumer supervisor (#3390)
A continuation of #3332 reducing the amount of memory used by the
consumer supervisor. Because supervisors hold the process and the config
needed to start it, the trick is to pass the absolute minimum required
config to the supervisor and have the process itself lookup anything
else it needs based on that.