fix: Match snapshot started timeout to expected time for snapshot to start (#3401)
Supersedes https://github.com/electric-sql/electric/pull/3391
Relevant to https://github.com/electric-sql/electric/issues/3365

Since the intent is to filter out cases of the snapshot just taking long
to start - using the default timeout and setting it to 45 seconds (which
is strictly larger than the 30 second timeout to start + potential
waiting time for publication updates and what not) so requests get held
for longer and should never time out.

I'm keeping the timeout on the call as a runtime error and thus a 500 so
we can monitor if we still get those, judging from the beam metrics my
suspicion is that the real culprit is the massive buildup of the message
queue in `ShapeLogCollector` due to subscribe call from consumers -
which means that we should still see time outs.

If we do indeed observe that, we either need to find a way to help SLC
handle these surges of shape adds (batch index updates? update the index
in a separate process, and copy over a diff?), or we can then make the
timeout return a proper error indicating the _our_ system is overloaded,
rather than the underlying PG.