fix: Avoid going through erl file server process for per-shape ops (#3403)
Started tracing calls to the file server with `:sys.trace` and running
Electric with new shapes being created and found a few spots we missed
that build up a queue on the file server.

If we want shapes to be scalable we need to avoid blocking on single
processes unless strictly necessary (e.g. currently for building the
index on the shape log collector)

Took the opportunity to remove unused code.